[groovy] spock: move to ext package, extract extensions to separate plugin.xml
