+++ plugins/groovy/src/org/jetbrains/plugins/groovy/ext/spock/SpockTestFramework.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jetbrains.plugins.groovy.spock;
+package org.jetbrains.plugins.groovy.ext.spock;
 
 import com.intellij.execution.junit.JUnitUtil;
 import com.intellij.ide.fileTemplates.FileTemplateDescriptor;
