+++ plugins/groovy/test/org/jetbrains/plugins/groovy/ext/spock/SpockTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2016 JetBrains s.r.o.
+ * Copyright 2000-2017 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jetbrains.plugins.groovy.spock
+package org.jetbrains.plugins.groovy.ext.spock
 
 import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.psi.CommonClassNames
